+++ kernels/HIP/fp16fp32/basic_matmul_col.cpp
@@ -6,7 +6,7 @@ extern "C" __global__ void matmul_kernel(int M, int N, int K, __half *A, __half 
     if (row < M && col < N) {
         float sum = 0.0f;
         for (int k = 0; k < K; k++) {
-            sum += __half2float(A[row * K + k] * B[k * N + col]);
+            sum += __half2float(A[row * K + k]) * __half2float(B[k * N + col]);
         }
         C[row * N + col] = alpha * sum + beta * C[row * N + col];
     }
